new pdf2ppt qwenvl
