+++ frontend-workflow/src/components/Paper2PptPage.tsx
@@ -1,1760 +1,3 @@
-import { useState, useEffect, ChangeEvent } from 'react';
-import { useTranslation } from 'react-i18next';
-import {
-  UploadCloud, Settings2, Download, Loader2, CheckCircle2,
-  AlertCircle, ChevronDown, ChevronUp, Github, Star, X, Sparkles,
-  ArrowRight, ArrowLeft, GripVertical, Trash2, Edit3, Check, RotateCcw,
-  MessageSquare, RefreshCw, FileText, Key, Globe, Cpu, Type, Lightbulb,
-  Copy, Share2, Info
-} from 'lucide-react';
-import { uploadAndSaveFile } from '../services/fileService';
-import { API_KEY } from '../config/api';
-import { checkQuota, recordUsage } from '../services/quotaService';
-import { verifyLlmConnection } from '../services/llmService';
-import { useAuthStore } from '../stores/authStore';
-import QRCodeTooltip from './QRCodeTooltip';
-
-// ============== 类型定义 ==============
-type Step = 'upload' | 'outline' | 'generate' | 'complete';
-
-interface SlideOutline {
-  id: string;
-  pageNum: number;
-  title: string;
-  layout_description: string;
-  key_points: string[];
-  asset_ref: string | null;
-}
-
-interface GenerateResult {
-  slideId: string;
-  beforeImage: string;
-  afterImage: string;
-  status: 'pending' | 'processing' | 'done';
-  userPrompt?: string;
-}
-
-const MAX_FILE_SIZE = 50 * 1024 * 1024; // 50MB
-
-// ============== 主组件 ==============
-const Paper2PptPage = () => {
-  const { t } = useTranslation(['paper2ppt', 'common']);
-  const { user, refreshQuota } = useAuthStore();
-  // Step 状态
-  const [currentStep, setCurrentStep] = useState<Step>('upload');
-  
-  // Step 1: 上传相关状态
-  const [uploadMode, setUploadMode] = useState<'file' | 'text' | 'topic'>('file');
-  const [textContent, setTextContent] = useState('');
-  const [selectedFile, setSelectedFile] = useState<File | null>(null);
-  const [isDragOver, setIsDragOver] = useState(false);
-  const [styleMode, setStyleMode] = useState<'prompt' | 'reference'>('prompt');
-  const [stylePreset, setStylePreset] = useState<'modern' | 'business' | 'academic' | 'creative'>('modern');
-  const [globalPrompt, setGlobalPrompt] = useState('');
-  const [referenceImage, setReferenceImage] = useState<File | null>(null);
-  const [referenceImagePreview, setReferenceImagePreview] = useState<string | null>(null);
-  const [isUploading, setIsUploading] = useState(false);
-  const [isValidating, setIsValidating] = useState(false);
-  const [pageCount, setPageCount] = useState(6);
-  const [useLongPaper, setUseLongPaper] = useState(false);
-  const [progress, setProgress] = useState(0);
-  const [progressStatus, setProgressStatus] = useState('');
-  
-  // Step 2: Outline 相关状态
-  const [outlineData, setOutlineData] = useState<SlideOutline[]>([]);
-  const [editingId, setEditingId] = useState<string | null>(null);
-  const [editContent, setEditContent] = useState<{
-    title: string;
-    layout_description: string;
-    key_points: string[];
-  }>({ title: '', layout_description: '', key_points: [] });
-  
-  // Step 3: 生成相关状态
-  const [currentSlideIndex, setCurrentSlideIndex] = useState(0);
-  const [generateResults, setGenerateResults] = useState<GenerateResult[]>([]);
-  const [isGenerating, setIsGenerating] = useState(false);
-  const [slidePrompt, setSlidePrompt] = useState('');
-  
-  // Step 4: 完成状态
-  const [isGeneratingFinal, setIsGeneratingFinal] = useState(false);
-  const [downloadUrl, setDownloadUrl] = useState<string | null>(null);
-  const [pdfPreviewUrl, setPdfPreviewUrl] = useState<string | null>(null);
-  
-  // 通用状态
-  const [error, setError] = useState<string | null>(null);
-  const [showBanner, setShowBanner] = useState(true);
-  
-  // API 配置状态
-  const [inviteCode, setInviteCode] = useState('');
-  const [llmApiUrl, setLlmApiUrl] = useState('https://api.apiyi.com/v1');
-  const [apiKey, setApiKey] = useState('');
-  const [model, setModel] = useState('gpt-5.1');
-  const [genFigModel, setGenFigModel] = useState('gemini-2.5-flash-image');
-  const [language, setLanguage] = useState<'zh' | 'en'>('en');
-  const [resultPath, setResultPath] = useState<string | null>(null);
-
-  // GitHub Stars
-  const [stars, setStars] = useState<{dataflow: number | null, agent: number | null, dataflex: number | null}>({
-    dataflow: null,
-    agent: null,
-    dataflex: null,
-  });
-  const [copySuccess, setCopySuccess] = useState('');
-
-  const shareText = `发现一个超好用的AI工具 DataFlow-Agent！🚀
-支持论文转PPT、PDF转PPT、PPT美化等功能，科研打工人的福音！
-
-🔗 在线体验：https://dcai-paper2any.nas.cpolar.cn/
-⭐ GitHub Agent：https://github.com/OpenDCAI/Paper2Any
-🌟 GitHub Core：https://github.com/OpenDCAI/DataFlow
-
-转发本文案+截图，联系微信群管理员即可获取免费Key！🎁
-#AI工具 #PPT制作 #科研效率 #开源项目`;
-
-  const handleCopyShareText = async () => {
-    try {
-      if (navigator.clipboard && window.isSecureContext) {
-        await navigator.clipboard.writeText(shareText);
-      } else {
-        const textArea = document.createElement("textarea");
-        textArea.value = shareText;
-        textArea.style.position = "fixed";
-        textArea.style.left = "-9999px";
-        textArea.style.top = "0";
-        document.body.appendChild(textArea);
-        textArea.focus();
-        textArea.select();
-        try {
-          document.execCommand('copy');
-        } catch (err) {
-          console.error('Fallback: Oops, unable to copy', err);
-          throw err;
-        } finally {
-          document.body.removeChild(textArea);
-        }
-      }
-      setCopySuccess('文案已复制！快去分享吧');
-      setTimeout(() => setCopySuccess(''), 2000);
-    } catch (err) {
-      console.error('复制失败', err);
-      setCopySuccess('复制失败，请手动复制');
-    }
-  };
-
-  useEffect(() => {
-    const fetchStars = async () => {
-      try {
-        const [res1, res2, res3] = await Promise.all([
-          fetch('https://api.github.com/repos/OpenDCAI/DataFlow'),
-          fetch('https://api.github.com/repos/OpenDCAI/Paper2Any'),
-          fetch('https://api.github.com/repos/OpenDCAI/DataFlex')
-        ]);
-        const data1 = await res1.json();
-        const data2 = await res2.json();
-        const data3 = await res3.json();
-        setStars({
-          dataflow: data1.stargazers_count,
-          agent: data2.stargazers_count,
-          dataflex: data3.stargazers_count,
-        });
-      } catch (e) {
-        console.error('Failed to fetch stars', e);
-      }
-    };
-    fetchStars();
-  }, []);
-
-  // ============== Step 1: 上传处理 ==============
-  const validateDocFile = (file: File): boolean => {
-    const ext = file.name.split('.').pop()?.toLowerCase();
-    if (ext !== 'pdf') {
-      setError('仅支持 PDF 格式');
-      return false;
-    }
-    return true;
-  };
-
-  const handleFileChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file || !validateDocFile(file)) return;
-    if (file.size > MAX_FILE_SIZE) {
-      setError('文件大小超过 50MB 限制');
-      return;
-    }
-    setSelectedFile(file);
-    setError(null);
-  };
-
-  const handleDrop = (e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    setIsDragOver(false);
-    const file = e.dataTransfer.files?.[0];
-    if (!file || !validateDocFile(file)) return;
-    if (file.size > MAX_FILE_SIZE) {
-      setError('文件大小超过 50MB 限制');
-      return;
-    }
-    setSelectedFile(file);
-    setError(null);
-  };
-
-  const handleReferenceImageChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    const ext = file.name.split('.').pop()?.toLowerCase();
-    if (!['jpg', 'jpeg', 'png', 'webp', 'gif'].includes(ext || '')) {
-      setError('参考图片仅支持 JPG/PNG/WEBP/GIF 格式');
-      return;
-    }
-    setReferenceImage(file);
-    setReferenceImagePreview(URL.createObjectURL(file));
-    setError(null);
-  };
-
-  const handleRemoveReferenceImage = () => {
-    if (referenceImagePreview) {
-      URL.revokeObjectURL(referenceImagePreview);
-    }
-    setReferenceImage(null);
-    setReferenceImagePreview(null);
-  };
-
-  const handleUploadAndParse = async () => {
-    if (uploadMode === 'file' && !selectedFile) {
-      setError('请先选择 PDF 文件');
-      return;
-    }
-    if ((uploadMode === 'text' || uploadMode === 'topic') && !textContent.trim()) {
-      setError(uploadMode === 'text' ? '请输入长文本内容' : '请输入 Topic 主题');
-      return;
-    }
-    
-    // if (!inviteCode.trim()) {
-    //   setError('请输入邀请码');
-    //   return;
-    // }
-    if (!apiKey.trim()) {
-      setError('请输入 API Key');
-      return;
-    }
-
-    // Check quota before proceeding
-    const quota = await checkQuota(user?.id || null, user?.is_anonymous || false);
-    if (quota.remaining <= 0) {
-      setError(quota.isAuthenticated
-        ? '今日配额已用完（10次/天），请明天再试'
-        : '今日配额已用完（5次/天），登录后可获得更多配额');
-      return;
-    }
-
-    try {
-        // Step 0: Verify LLM Connection first
-        setIsValidating(true);
-        setError(null);
-        await verifyLlmConnection(llmApiUrl, apiKey, model);
-        setIsValidating(false);
-    } catch (err) {
-        setIsValidating(false);
-        const message = err instanceof Error ? err.message : 'API 验证失败';
-        setError(message);
-        return; // Stop execution if validation fails
-    }
-
-    setIsUploading(true);
-    setError(null);
-    setProgress(0);
-    setProgressStatus('正在初始化...');
-    
-    // 模拟进度
-    const progressInterval = setInterval(() => {
-      setProgress(prev => {
-        if (prev >= 90) return 90;
-        const messages = [
-           '正在内容准备...',
-           '正在解析内容...',
-           '正在分析结构...',
-           '正在提取关键点...',
-           '正在生成大纲...'
-        ];
-        const msgIndex = Math.floor(prev / 20);
-        if (msgIndex < messages.length) {
-          setProgressStatus(messages[msgIndex]);
-        }
-        // 调整进度速度，使其在 3 分钟左右达到 90%
-        return prev + (Math.random() * 0.6 + 0.2);
-      });
-    }, 1000);
-
-    try {
-      const formData = new FormData();
-      if (uploadMode === 'file' && selectedFile) {
-        formData.append('file', selectedFile);
-        formData.append('input_type', 'pdf');
-      } else {
-        formData.append('text', textContent.trim());
-        formData.append('input_type', uploadMode); // 'text' or 'topic'
-      }
-      
-      formData.append('invite_code', inviteCode.trim());
-      formData.append('chat_api_url', llmApiUrl.trim());
-      formData.append('api_key', apiKey.trim());
-      formData.append('model', model);
-      formData.append('language', language);
-      formData.append('style', globalPrompt || getStyleDescription(stylePreset));
-      formData.append('gen_fig_model', genFigModel);
-      formData.append('page_count', String(pageCount));
-      formData.append('use_long_paper', String(useLongPaper));
-
-      if (styleMode === 'reference' && referenceImage) {
-        formData.append('reference_img', referenceImage);
-        // 如果有参考图，清空 style 参数，避免混淆
-        formData.set('style', '');
-      }
-      
-      console.log(`Sending request to /api/paper2ppt/pagecontent_json with input_type=${uploadMode}`);
-      
-      const res = await fetch('/api/paper2ppt/pagecontent_json', {
-        method: 'POST',
-        headers: { 'X-API-Key': API_KEY },
-        body: formData,
-      });
-      
-      if (!res.ok) {
-        let msg = '服务器繁忙，请稍后再试';
-        if (res.status === 403) {
-          msg = '邀请码不正确或已失效';
-        } else if (res.status === 429) {
-          msg = '请求过于频繁，请稍后再试';
-        }
-        throw new Error(msg);
-      }
-      
-      const data = await res.json();
-      console.log('API Response:', JSON.stringify(data, null, 2));
-      
-      if (!data.success) {
-        throw new Error('服务器繁忙，请稍后再试');
-      }
-      
-      const currentResultPath = data.result_path || '';
-      if (currentResultPath) {
-        setResultPath(currentResultPath);
-      } else {
-        throw new Error('后端未返回 result_path');
-      }
-      
-      if (!data.pagecontent || data.pagecontent.length === 0) {
-        throw new Error('解析结果为空，请检查输入内容是否正确');
-      }
-      
-      const convertedSlides: SlideOutline[] = data.pagecontent.map((item: any, index: number) => ({
-        id: String(index + 1),
-        pageNum: index + 1,
-        title: item.title || `第 ${index + 1} 页`,
-        layout_description: item.layout_description || '',
-        key_points: item.key_points || [],
-        asset_ref: item.asset_ref || null,
-      }));
-      
-      clearInterval(progressInterval);
-      setProgress(100);
-      setProgressStatus('解析完成！');
-      
-      // 稍微延迟一下跳转，让用户看到 100%
-      setTimeout(() => {
-        setOutlineData(convertedSlides);
-        setCurrentStep('outline');
-      }, 500);
-      
-    } catch (err) {
-      clearInterval(progressInterval);
-      setProgress(0);
-      const message = err instanceof Error ? err.message : '服务器繁忙，请稍后再试';
-      setError(message);
-      console.error(err);
-    } finally {
-      if (currentStep !== 'outline') {
-         setIsUploading(false);
-      } else {
-         // 如果成功跳转，在组件卸载或状态切换前保持 loading 状态防止闪烁，这里不需要特别处理，因为 setCurrentStep 已经切换了视图
-         setIsUploading(false);
-      }
-    }
-  };
-
-  const getStyleDescription = (preset: string): string => {
-    // Note: These descriptions are sent to the backend, so they might need to stay in the language the backend expects (likely Chinese or English depending on prompt).
-    // If the backend uses these for image generation prompts, English is usually better, or keep Chinese if the model is optimized for it.
-    // For now, I will keep them hardcoded or use t() if they are just for display.
-    // But wait, these are used in `formData.append('style', ...)`
-    // If I translate them, the backend receives translated strings.
-    // Assuming backend handles it or LLM handles it.
-    // Let's use t() but ensure keys exist. I didn't add these specific long descriptions to JSON.
-    // I'll leave them as is for now to avoid breaking generation logic, or map them if I added them.
-    // I added "presets" keys but not the descriptions.
-    // Let's keep them hardcoded for safety as they are prompt inputs, unless I add them to JSON.
-    // Actually, let's leave them hardcoded for now as they are "backend values".
-    const styles: Record<string, string> = {
-      modern: '现代简约风格，使用干净的线条和充足的留白',
-      business: '商务专业风格，稳重大气，适合企业演示',
-      academic: '学术报告风格，清晰的层次结构，适合论文汇报',
-      creative: '创意设计风格，活泼生动，色彩丰富',
-    };
-    return styles[preset] || styles.modern;
-  };
-
-  // ============== Step 2: Outline 编辑处理 ==============
-  const handleEditStart = (slide: SlideOutline) => {
-    setEditingId(slide.id);
-    setEditContent({ 
-      title: slide.title, 
-      layout_description: slide.layout_description,
-      key_points: [...slide.key_points]
-    });
-  };
-
-  const handleEditSave = () => {
-    if (!editingId) return;
-    setOutlineData(prev => prev.map(s => 
-      s.id === editingId 
-        ? { ...s, title: editContent.title, layout_description: editContent.layout_description, key_points: editContent.key_points }
-        : s
-    ));
-    setEditingId(null);
-  };
-
-  const handleKeyPointChange = (index: number, value: string) => {
-    setEditContent(prev => {
-      const newKeyPoints = [...prev.key_points];
-      newKeyPoints[index] = value;
-      return { ...prev, key_points: newKeyPoints };
-    });
-  };
-
-  const handleAddKeyPoint = () => {
-    setEditContent(prev => ({ ...prev, key_points: [...prev.key_points, ''] }));
-  };
-
-  const handleRemoveKeyPoint = (index: number) => {
-    setEditContent(prev => ({ ...prev, key_points: prev.key_points.filter((_, i) => i !== index) }));
-  };
-
-  const handleEditCancel = () => setEditingId(null);
-  
-  const handleDeleteSlide = (id: string) => {
-    setOutlineData(prev => prev.filter(s => s.id !== id).map((s, i) => ({ ...s, pageNum: i + 1, id: String(i + 1) })));
-  };
-  
-  const handleMoveSlide = (index: number, direction: 'up' | 'down') => {
-    const newData = [...outlineData];
-    const targetIndex = direction === 'up' ? index - 1 : index + 1;
-    if (targetIndex < 0 || targetIndex >= newData.length) return;
-    [newData[index], newData[targetIndex]] = [newData[targetIndex], newData[index]];
-    setOutlineData(newData.map((s, i) => ({ ...s, pageNum: i + 1 })));
-  };
-
-  const handleConfirmOutline = async () => {
-    setCurrentStep('generate');
-    setCurrentSlideIndex(0);
-    setIsGenerating(true);
-    setError(null);
-    
-    const results: GenerateResult[] = outlineData.map((slide) => ({
-      slideId: slide.id,
-      beforeImage: '',
-      afterImage: '',
-      status: 'processing' as const,
-    }));
-    setGenerateResults(results);
-    
-    try {
-      const formData = new FormData();
-      formData.append('img_gen_model_name', genFigModel);
-      formData.append('chat_api_url', llmApiUrl.trim());
-      formData.append('api_key', apiKey.trim());
-      formData.append('model', model);
-      formData.append('language', language);
-      formData.append('style', globalPrompt || getStyleDescription(stylePreset));
-      formData.append('aspect_ratio', '16:9');
-      formData.append('invite_code', inviteCode.trim());
-      formData.append('result_path', resultPath || '');
-      formData.append('get_down', 'false');
-
-      const pagecontent = outlineData.map((slide) => ({
-        title: slide.title,
-        layout_description: slide.layout_description,
-        key_points: slide.key_points,
-        asset_ref: slide.asset_ref,
-      }));
-      formData.append('pagecontent', JSON.stringify(pagecontent));
-
-      const res = await fetch('/api/paper2ppt/ppt_json', {
-        method: 'POST',
-        headers: { 'X-API-Key': API_KEY },
-        body: formData,
-      });
-      
-      if (!res.ok) {
-        let msg = '服务器繁忙，请稍后再试';
-        if (res.status === 429) {
-          msg = '请求过于频繁，请稍后再试';
-        }
-        throw new Error(msg);
-      }
-      
-      const data = await res.json();
-      
-      if (!data.success) {
-        throw new Error('服务器繁忙，请稍后再试');
-      }
-      
-      const updatedResults = results.map((result, index) => {
-        const pageNumStr = String(index).padStart(3, '0');
-        let afterImage = '';
-        
-        if (data.all_output_files && Array.isArray(data.all_output_files)) {
-          const pageImg = data.all_output_files.find((url: string) => 
-            url.includes(`ppt_pages/page_${pageNumStr}.png`)
-          );
-          if (pageImg) {
-            afterImage = pageImg;
-          }
-        }
-        
-        return {
-          ...result,
-          afterImage,
-          status: 'done' as const,
-        };
-      });
-      
-      // 预加载所有图片到浏览器缓存，避免切换页面时延迟
-      if (data.all_output_files && Array.isArray(data.all_output_files)) {
-        console.log('预加载所有生成的图片...');
-        data.all_output_files.forEach((url: string) => {
-          if (url.endsWith('.png') || url.endsWith('.jpg') || url.endsWith('.jpeg')) {
-            const img = new Image();
-            img.src = url;
-          }
-        });
-      }
-      
-      setGenerateResults(updatedResults);
-      
-    } catch (err) {
-      const message = err instanceof Error ? err.message : '服务器繁忙，请稍后再试';
-      setError(message);
-      setGenerateResults(results.map(r => ({ ...r, status: 'pending' as const })));
-    } finally {
-      setIsGenerating(false);
-    }
-  };
-
-  // ============== Step 3: 重新生成单页 ==============
-  const handleRegenerateSlide = async () => {
-    if (!resultPath) {
-      setError('缺少 result_path，请重新上传文件');
-      return;
-    }
-    
-    if (!slidePrompt.trim()) {
-      setError('请输入重新生成的提示词');
-      return;
-    }
-    
-    setIsGenerating(true);
-    setError(null);
-    
-    const updatedResults = [...generateResults];
-    updatedResults[currentSlideIndex] = { 
-      ...updatedResults[currentSlideIndex], 
-      status: 'processing',
-      userPrompt: slidePrompt,
-    };
-    setGenerateResults(updatedResults);
-    
-    try {
-      const formData = new FormData();
-      formData.append('img_gen_model_name', genFigModel);
-      formData.append('chat_api_url', llmApiUrl.trim());
-      formData.append('api_key', apiKey.trim());
-      formData.append('model', model);
-      formData.append('language', language);
-      formData.append('style', globalPrompt || getStyleDescription(stylePreset));
-      formData.append('aspect_ratio', '16:9');
-      formData.append('invite_code', inviteCode.trim());
-      formData.append('result_path', resultPath);
-      formData.append('get_down', 'true');
-      formData.append('page_id', String(currentSlideIndex));
-      formData.append('edit_prompt', slidePrompt);
-
-      const pagecontent = outlineData.map((slide, idx) => {
-        const result = generateResults[idx];
-        let generatedPath = '';
-        if (result?.afterImage) {
-          // 直接使用 URL，后端会自动处理
-          generatedPath = result.afterImage;
-        }
-        return {
-          title: slide.title,
-          layout_description: slide.layout_description,
-          key_points: slide.key_points,
-          asset_ref: slide.asset_ref,
-          generated_img_path: generatedPath || undefined,
-        };
-      });
-      formData.append('pagecontent', JSON.stringify(pagecontent));
-
-      const res = await fetch('/api/paper2ppt/ppt_json', {
-        method: 'POST',
-        headers: { 'X-API-Key': API_KEY },
-        body: formData,
-      });
-      
-      if (!res.ok) {
-        let msg = '服务器繁忙，请稍后再试';
-        if (res.status === 429) {
-          msg = '请求过于频繁，请稍后再试';
-        }
-        throw new Error(msg);
-      }
-      
-      const data = await res.json();
-      
-      if (!data.success) {
-        throw new Error('服务器繁忙，请稍后再试');
-      }
-      
-      const pageNumStr = String(currentSlideIndex).padStart(3, '0');
-      let afterImage = updatedResults[currentSlideIndex].afterImage;
-      
-      if (data.all_output_files && Array.isArray(data.all_output_files)) {
-        const pageImg = data.all_output_files.find((url: string) => 
-          url.includes(`ppt_pages/page_${pageNumStr}.png`)
-        );
-        if (pageImg) {
-          afterImage = pageImg + '?t=' + Date.now();
-        }
-      }
-      
-      updatedResults[currentSlideIndex] = { 
-        ...updatedResults[currentSlideIndex], 
-        afterImage,
-        status: 'done',
-      };
-      setGenerateResults([...updatedResults]);
-      setSlidePrompt('');
-      
-    } catch (err) {
-      const message = err instanceof Error ? err.message : '服务器繁忙，请稍后再试';
-      setError(message);
-      updatedResults[currentSlideIndex] = { 
-        ...updatedResults[currentSlideIndex], 
-        status: 'done',
-      };
-      setGenerateResults([...updatedResults]);
-    } finally {
-      setIsGenerating(false);
-    }
-  };
-
-  const handleConfirmSlide = () => {
-    setError(null);
-    if (currentSlideIndex < outlineData.length - 1) {
-      const nextIndex = currentSlideIndex + 1;
-      setCurrentSlideIndex(nextIndex);
-      setSlidePrompt('');
-    } else {
-      setCurrentStep('complete');
-    }
-  };
-
-  // ============== Step 4: 完成处理 ==============
-  const handleGenerateFinal = async () => {
-    if (!resultPath) {
-      setError('缺少 result_path');
-      return;
-    }
-    
-    setIsGeneratingFinal(true);
-    setError(null);
-    
-    try {
-      const formData = new FormData();
-      formData.append('img_gen_model_name', genFigModel);
-      formData.append('chat_api_url', llmApiUrl.trim());
-      formData.append('api_key', apiKey.trim());
-      formData.append('model', model);
-      formData.append('language', language);
-      formData.append('style', globalPrompt || getStyleDescription(stylePreset));
-      formData.append('aspect_ratio', '16:9');
-      formData.append('invite_code', inviteCode.trim());
-      formData.append('result_path', resultPath);
-      formData.append('get_down', 'false');
-      formData.append('all_edited_down', 'true');
-
-      const pagecontent = outlineData.map((slide) => ({
-        title: slide.title,
-        layout_description: slide.layout_description,
-        key_points: slide.key_points,
-        asset_ref: slide.asset_ref,
-      }));
-      formData.append('pagecontent', JSON.stringify(pagecontent));
-
-      const res = await fetch('/api/paper2ppt/ppt_json', {
-        method: 'POST',
-        headers: { 'X-API-Key': API_KEY },
-        body: formData,
-      });
-      
-      if (!res.ok) {
-        let msg = '服务器繁忙，请稍后再试';
-        if (res.status === 429) {
-          msg = '请求过于频繁，请稍后再试';
-        }
-        throw new Error(msg);
-      }
-      
-      const data = await res.json();
-      
-      if (!data.success) {
-        throw new Error('服务器繁忙，请稍后再试');
-      }
-      
-      // 优先使用后端直接返回的路径
-      if (data.ppt_pptx_path) {
-        setDownloadUrl(data.ppt_pptx_path);
-      }
-      if (data.ppt_pdf_path) {
-        setPdfPreviewUrl(data.ppt_pdf_path);
-      }
-      
-      // 备选：从 all_output_files 中查找
-      if (data.all_output_files && Array.isArray(data.all_output_files)) {
-        if (!data.ppt_pptx_path) {
-          const pptxFile = data.all_output_files.find((url: string) => 
-            url.endsWith('.pptx') || url.includes('editable.pptx')
-          );
-          if (pptxFile) {
-            setDownloadUrl(pptxFile);
-          }
-        }
-        if (!data.ppt_pdf_path) {
-          const pdfFile = data.all_output_files.find((url: string) =>
-            url.endsWith('.pdf') && !url.includes('input')
-          );
-          if (pdfFile) {
-            setPdfPreviewUrl(pdfFile);
-          }
-        }
-      }
-
-      // Record usage
-      await recordUsage(user?.id || null, 'paper2ppt');
-      refreshQuota();
-
-      // Upload generated file to Supabase Storage (either PPTX or PDF)
-      // Find PPTX file first (preferred)
-      let filePath = data.ppt_pptx_path || (data.all_output_files?.find((url: string) =>
-        url.endsWith('.pptx') || url.includes('editable.pptx')
-      ));
-      let defaultName = 'paper2ppt_result.pptx';
-
-      // If no PPTX, try PDF (exclude input PDFs)
-      if (!filePath) {
-        filePath = data.ppt_pdf_path || (data.all_output_files?.find((url: string) =>
-          url.endsWith('.pdf') && !url.includes('input')
-        ));
-        defaultName = 'paper2ppt_result.pdf';
-      }
-
-      if (filePath) {
-        try {
-          // Fix Mixed Content issue: upgrade http to https if current page is https
-          let fetchUrl = filePath;
-          if (window.location.protocol === 'https:' && filePath.startsWith('http:')) {
-            fetchUrl = filePath.replace('http:', 'https:');
-          }
-
-          const fileRes = await fetch(fetchUrl);
-          if (fileRes.ok) {
-            const fileBlob = await fileRes.blob();
-            const fileName = filePath.split('/').pop() || defaultName;
-            console.log('[Paper2PptPage] Uploading file to storage:', fileName);
-            await uploadAndSaveFile(fileBlob, fileName, 'paper2ppt');
-            console.log('[Paper2PptPage] File uploaded successfully');
-          } else {
-             console.error('[Paper2PptPage] Failed to fetch file for upload:', fileRes.status, fileRes.statusText);
-          }
-        } catch (e) {
-          console.error('[Paper2PptPage] Failed to upload file:', e);
-        }
-      }
-
-    } catch (err) {
-      const message = err instanceof Error ? err.message : '服务器繁忙，请稍后再试';
-      setError(message);
-    } finally {
-      setIsGeneratingFinal(false);
-    }
-  };
-
-  const handleDownload = () => {
-    if (!downloadUrl) return;
-    const a = document.createElement('a');
-    a.href = downloadUrl;
-    a.download = 'paper2ppt_result.pptx';
-    a.click();
-  };
-
-  const handleDownloadPdf = () => {
-    if (!pdfPreviewUrl) return;
-    // 在新窗口打开 PDF
-    window.open(pdfPreviewUrl, '_blank');
-  };
-
-  // ============== 渲染函数 ==============
-  const renderStepIndicator = () => {
-    const steps = [
-      { key: 'upload', label: t('steps.upload'), num: 1 },
-      { key: 'outline', label: t('steps.outline'), num: 2 },
-      { key: 'generate', label: t('steps.generate'), num: 3 },
-      { key: 'complete', label: t('steps.complete'), num: 4 },
-    ];
-    const currentIndex = steps.findIndex(s => s.key === currentStep);
-    return (
-      <div className="flex items-center justify-center gap-2 mb-8">
-        {steps.map((step, index) => (
-          <div key={step.key} className="flex items-center">
-            <div className={`flex items-center gap-2 px-4 py-2 rounded-full text-sm font-medium transition-all ${
-              index === currentIndex 
-                ? 'bg-gradient-to-r from-purple-500 to-pink-500 text-white shadow-lg' 
-                : index < currentIndex 
-                  ? 'bg-purple-500/20 text-purple-300 border border-purple-500/40' 
-                  : 'bg-white/5 text-gray-500 border border-white/10'
-            }`}>
-              <span className={`w-6 h-6 rounded-full flex items-center justify-center text-xs ${
-                index < currentIndex ? 'bg-purple-400 text-white' : ''
-              }`}>
-                {index < currentIndex ? <Check size={14} /> : step.num}
-              </span>
-              <span className="hidden sm:inline">{step.label}</span>
-            </div>
-            {index < steps.length - 1 && (
-              <ArrowRight size={16} className={`mx-2 ${index < currentIndex ? 'text-purple-400' : 'text-gray-600'}`} />
-            )}
-          </div>
-        ))}
-      </div>
-    );
-  };
-
-  const renderUploadStep = () => (
-    <div className="max-w-6xl mx-auto">
-      <div className="mb-10 text-center">
-        <p className="text-xs uppercase tracking-[0.2em] text-purple-300 mb-3 font-semibold">{t('upload.subtitle')}</p>
-        <h1 className="text-4xl md:text-5xl font-bold mb-4">
-          <span className="bg-gradient-to-r from-purple-400 via-pink-400 to-rose-400 bg-clip-text text-transparent">
-            {t('upload.title')}
-          </span>
-        </h1>
-        <p className="text-base text-gray-300 max-w-2xl mx-auto leading-relaxed">
-          {t('upload.desc')}<br />
-          <span className="text-purple-400">{t('upload.descHighlight')}</span>
-        </p>
-      </div>
-
-      <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-        {/* 左侧：输入区域 */}
-        <div className="glass rounded-xl border border-white/10 p-6 relative overflow-hidden">
-          {/* 装饰背景光 */}
-          <div className="absolute top-0 left-1/2 -translate-x-1/2 w-2/3 h-1 bg-gradient-to-r from-transparent via-purple-500 to-transparent opacity-50 blur-sm"></div>
-
-          {/* 炫酷模式切换 Tabs */}
-          <div className="grid grid-cols-3 gap-3 mb-6 p-1.5 bg-black/40 rounded-2xl border border-white/5">
-            {[
-              { id: 'file', label: t('upload.tabs.file'), icon: FileText, sub: t('upload.tabs.fileSub') },
-              { id: 'text', label: t('upload.tabs.text'), icon: Type, sub: t('upload.tabs.textSub') },
-              { id: 'topic', label: t('upload.tabs.topic'), icon: Lightbulb, sub: t('upload.tabs.topicSub') },
-            ].map((item) => (
-              <button 
-                key={item.id}
-                onClick={() => setUploadMode(item.id as any)}
-                className={`relative group flex flex-col items-center justify-center py-3 rounded-xl transition-all duration-300 overflow-hidden ${
-                  uploadMode === item.id 
-                    ? 'bg-gradient-to-br from-purple-600 to-pink-600 text-white shadow-lg shadow-purple-500/30 scale-[1.02] ring-1 ring-white/20' 
-                    : 'bg-white/5 text-gray-400 hover:bg-white/10 hover:text-gray-200 hover:scale-[1.02]'
-                }`}
-              >
-                {/* 选中态的光效扫光动画 */}
-                {uploadMode === item.id && (
-                  <div className="absolute inset-0 w-full h-full bg-gradient-to-r from-transparent via-white/20 to-transparent -translate-x-full animate-shimmer-fast"></div>
-                )}
-                
-                <item.icon size={22} className={`mb-1.5 transition-colors ${uploadMode === item.id ? 'text-white' : 'text-gray-500 group-hover:text-purple-400'}`} />
-                <span className={`text-sm font-bold tracking-wide ${uploadMode === item.id ? 'text-white' : 'text-gray-300'}`}>{item.label}</span>
-                <span className={`text-[10px] uppercase tracking-wider font-medium ${uploadMode === item.id ? 'text-purple-100' : 'text-gray-600'}`}>{item.sub}</span>
-              </button>
-            ))}
-          </div>
-
-          <div className="mb-3 flex items-center gap-2 px-1">
-            <span className="w-1 h-4 rounded-full bg-purple-500"></span>
-            <h3 className="text-white font-medium text-sm">
-              {uploadMode === 'file' ? t('upload.instruction.file') : uploadMode === 'text' ? t('upload.instruction.text') : t('upload.instruction.topic')}
-            </h3>
-          </div>
-
-          {uploadMode === 'file' ? (
-            <div 
-              className={`border-2 border-dashed rounded-xl p-8 flex flex-col items-center justify-center text-center gap-4 transition-all h-[300px] ${
-                isDragOver ? 'border-purple-500 bg-purple-500/10' : 'border-white/20 hover:border-purple-400'
-              }`} 
-              onDragOver={e => { e.preventDefault(); setIsDragOver(true); }} 
-              onDragLeave={e => { e.preventDefault(); setIsDragOver(false); }} 
-              onDrop={handleDrop}
-            >
-              <div className="w-16 h-16 rounded-full bg-gradient-to-br from-purple-500/20 to-pink-500/20 flex items-center justify-center">
-                <UploadCloud size={32} className="text-purple-400" />
-              </div>
-              <div>
-                <p className="text-white font-medium mb-1">{t('upload.dropzone.dragText')}</p>
-                <p className="text-sm text-gray-400">{t('upload.dropzone.supportText')}</p>
-              </div>
-              <label className="px-6 py-2.5 rounded-full bg-gradient-to-r from-purple-600 to-pink-600 text-white text-sm font-medium cursor-pointer hover:from-purple-700 hover:to-pink-700 transition-all">
-                {t('upload.dropzone.button')}
-                <input type="file" accept=".pdf" className="hidden" onChange={handleFileChange} />
-              </label>
-              {selectedFile && (
-                <div className="px-4 py-2 bg-purple-500/20 border border-purple-500/40 rounded-lg">
-                  <p className="text-sm text-purple-300">✓ {selectedFile.name}</p>
-                  <p className="text-xs text-gray-400 mt-1">✨ {t('upload.dropzone.analyzing')}</p>
-                </div>
-              )}
-            </div>
-          ) : (
-            <div className="flex flex-col h-[300px]">
-              <textarea
-                value={textContent}
-                onChange={e => setTextContent(e.target.value)}
-                placeholder={uploadMode === 'text' 
-                  ? t('upload.textInput.placeholderText')
-                  : t('upload.textInput.placeholderTopic')}
-                className="flex-1 w-full rounded-xl border border-white/20 bg-black/40 px-4 py-3 text-sm text-gray-100 outline-none focus:ring-2 focus:ring-purple-500 resize-none"
-              />
-              <p className="text-xs text-gray-500 mt-2 text-right">
-                {uploadMode === 'text' ? `${textContent.length} ${t('upload.textInput.charCount')}` : t('upload.textInput.deepResearch')}
-              </p>
-            </div>
-          )}
-        </div>
-
-        {/* 右侧：配置区域 */}
-        <div className="glass rounded-xl border border-white/10 p-6 space-y-4">
-          <h3 className="text-white font-semibold flex items-center gap-2">
-            <Settings2 size={18} className="text-purple-400" /> {t('upload.config.title')}
-          </h3>
-          
-          {/* API 配置 */}
-          <div className="grid grid-cols-2 gap-3">
-            <div>
-              <label className="block text-xs text-gray-400 mb-1 flex items-center gap-1">
-                <Key size={12} /> {t('upload.config.apiKey')}
-              </label>
-              <input 
-                type="password" 
-                value={apiKey} 
-                onChange={e => setApiKey(e.target.value)}
-                placeholder={t('upload.config.apiKeyPlaceholder')}
-                className="w-full rounded-lg border border-white/20 bg-black/40 px-3 py-2 text-sm text-gray-100 outline-none focus:ring-2 focus:ring-purple-500"
-              />
-            </div>
-          </div>
-          
-          <div className="grid grid-cols-2 gap-3">
-            <div>
-              <div className="flex items-center justify-between mb-1">
-                <label className="block text-xs text-gray-400 flex items-center gap-1">
-                  <Globe size={12} /> {t('upload.config.apiUrl')}
-                </label>
-                <QRCodeTooltip>
-                <a
-                  href={llmApiUrl === 'http://123.129.219.111:3000/v1' ? "http://123.129.219.111:3000" : "https://api.apiyi.com/register/?aff_code=TbrD"}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="text-[10px] text-purple-300 hover:text-purple-200 hover:underline"
-                >
-                  {t('upload.config.buyLink')}
-                </a>
-                </QRCodeTooltip>
-              </div>
-              <select 
-                value={llmApiUrl} 
-                onChange={e => {
-                  const val = e.target.value;
-                  setLlmApiUrl(val);
-                  if (val === 'http://123.129.219.111:3000/v1') {
-                    setGenFigModel('gemini-3-pro-image-preview');
-                  }
-                }}
-                className="w-full rounded-lg border border-white/20 bg-black/40 px-3 py-2 text-sm text-gray-100 outline-none focus:ring-2 focus:ring-purple-500"
-              >
-                <option value="https://api.apiyi.com/v1">https://api.apiyi.com/v1</option>
-                <option value="http://b.apiyi.com:16888/v1">http://b.apiyi.com:16888/v1</option>
-                <option value="http://123.129.219.111:3000/v1">http://123.129.219.111:3000/v1</option>
-              </select>
-            </div>
-            <div>
-              <label className="block text-xs text-gray-400 mb-1 flex items-center gap-1">
-                <Cpu size={12} /> {t('upload.config.model')}
-              </label>
-              <select 
-                value={model} 
-                onChange={e => setModel(e.target.value)}
-                className="w-full rounded-lg border border-white/20 bg-black/40 px-3 py-2 text-sm text-gray-100 outline-none focus:ring-2 focus:ring-purple-500"
-              >
-                <option value="gpt-4o">gpt-4o</option>
-                <option value="gpt-5.1">gpt-5.1</option>
-                <option value="gpt-5.2">gpt-5.2</option>
-                <option value="gemini-3-pro-preview">gemini-3-pro-preview</option>
-              </select>
-            </div>
-          </div>
-          
-          <div className="grid grid-cols-2 gap-3">
-            <div>
-              <label className="block text-xs text-gray-400 mb-1">{t('upload.config.genModel')}</label>
-              <select
-                value={genFigModel}
-                onChange={e => setGenFigModel(e.target.value)}
-                disabled={llmApiUrl === 'http://123.129.219.111:3000/v1'}
-                className="w-full rounded-lg border border-white/20 bg-black/40 px-3 py-2 text-sm text-gray-100 outline-none focus:ring-2 focus:ring-purple-500 disabled:opacity-50 disabled:cursor-not-allowed"
-              >
-                <option value="gemini-2.5-flash-image">Gemini 2.5 (Flash Image)</option>
-                <option value="gemini-3-pro-image-preview">Gemini 3 Pro (中文推荐)</option>
-              </select>
-              {llmApiUrl === 'http://123.129.219.111:3000/v1' && (
-                 <p className="text-[10px] text-gray-500 mt-1">此源仅支持 gemini-3-pro</p>
-              )}
-            </div>
-            <div>
-              <label className="block text-xs text-gray-400 mb-1">{t('upload.config.pageCount')}</label>
-              <input 
-                type="number" 
-                value={pageCount} 
-                onChange={e => setPageCount(parseInt(e.target.value) || 6)}
-                min={1}
-                max={20}
-                className="w-full rounded-lg border border-white/20 bg-black/40 px-3 py-2 text-sm text-gray-100 outline-none focus:ring-2 focus:ring-purple-500"
-              />
-            </div>
-          </div>
-
-          <div className="flex items-center gap-2 px-1 py-1">
-            <button
-              onClick={() => setUseLongPaper(!useLongPaper)}
-              className={`relative inline-flex h-5 w-9 items-center rounded-full transition-colors ${
-                useLongPaper ? 'bg-purple-600' : 'bg-gray-600'
-              }`}
-            >
-              <span
-                className={`inline-block h-3 w-3 transform rounded-full bg-white transition-transform ${
-                  useLongPaper ? 'translate-x-5' : 'translate-x-1'
-                }`}
-              />
-            </button>
-            <span className="text-xs text-gray-300 cursor-pointer" onClick={() => setUseLongPaper(!useLongPaper)}>
-              {t('upload.config.longPaper')}
-            </span>
-          </div>
-
-          <div className="border-t border-white/10 pt-4 mt-2">
-            <h4 className="text-xs text-gray-400 mb-2">{t('upload.config.styleTitle')}</h4>
-            
-            <div className="mb-3">
-              <label className="block text-xs text-gray-400 mb-1">{t('upload.config.language')}</label>
-              <select 
-                value={language} 
-                onChange={e => setLanguage(e.target.value as 'zh' | 'en')} 
-                className="w-full rounded-lg border border-white/20 bg-black/40 px-3 py-2 text-sm text-gray-100 outline-none focus:ring-2 focus:ring-purple-500"
-              >
-                <option value="zh">中文</option>
-                <option value="en">English</option>
-              </select>
-            </div>
-
-            <div className="flex gap-2 mb-3">
-              <button
-                type="button"
-                onClick={() => setStyleMode('prompt')}
-                className={`flex-1 py-2.5 px-3 rounded-lg text-xs font-medium flex items-center justify-center gap-1 transition-all ${
-                  styleMode === 'prompt'
-                    ? 'bg-gradient-to-r from-purple-500 to-pink-500 text-white shadow-sm'
-                    : 'bg-white/5 text-gray-400 border border-white/10 hover:bg-white/10'
-                }`}
-              >
-                <Sparkles size={14} /> {t('upload.config.styleMode.prompt')}
-              </button>
-              <button
-                type="button"
-                onClick={() => setStyleMode('reference')}
-                className={`flex-1 py-2.5 px-3 rounded-lg text-xs font-medium flex items-center justify-center gap-1 transition-all ${
-                  styleMode === 'reference'
-                    ? 'bg-gradient-to-r from-purple-500 to-pink-500 text-white shadow-sm'
-                    : 'bg-white/5 text-gray-400 border border-white/10 hover:bg-white/10'
-                }`}
-              >
-                <UploadCloud size={14} /> {t('upload.config.styleMode.reference')}
-              </button>
-            </div>
-
-            {styleMode === 'prompt' ? (
-              <>
-                <div className="mb-3">
-                  <label className="block text-xs text-gray-400 mb-1">{t('upload.config.stylePreset')}</label>
-                  <select 
-                    value={stylePreset} 
-                    onChange={e => setStylePreset(e.target.value as typeof stylePreset)} 
-                    className="w-full rounded-lg border border-white/20 bg-black/40 px-3 py-2 text-sm text-gray-100 outline-none focus:ring-2 focus:ring-purple-500"
-                  >
-                    <option value="modern">{t('upload.config.presets.modern')}</option>
-                    <option value="business">{t('upload.config.presets.business')}</option>
-                    <option value="academic">{t('upload.config.presets.academic')}</option>
-                    <option value="creative">{t('upload.config.presets.creative')}</option>
-                  </select>
-                </div>
-                <div>
-                  <label className="block text-xs text-gray-400 mb-1">{t('upload.config.promptLabel')}</label>
-                  <textarea 
-                    value={globalPrompt} 
-                    onChange={e => setGlobalPrompt(e.target.value)} 
-                    placeholder={t('upload.config.promptPlaceholder')}
-                    rows={2} 
-                    className="w-full rounded-lg border border-white/20 bg-black/40 px-3 py-2 text-sm text-gray-100 outline-none focus:ring-2 focus:ring-purple-500 resize-none" 
-                  />
-                </div>
-              </>
-            ) : (
-              <div>
-                <label className="block text-xs text-gray-400 mb-1">{t('upload.config.referenceLabel')}</label>
-                {referenceImagePreview ? (
-                  <div className="relative">
-                    <img
-                      src={referenceImagePreview}
-                      alt="参考风格"
-                      className="w-full h-32 object-cover rounded-lg border border-white/20"
-                    />
-                    <button
-                      type="button"
-                      onClick={handleRemoveReferenceImage}
-                      className="absolute top-2 right-2 p-1.5 rounded-full bg-black/60 text-white hover:bg-red-500 transition-colors"
-                    >
-                      <X size={14} />
-                    </button>
-                    <p className="text-[11px] text-purple-300 mt-1">✓ {t('upload.config.referenceUploaded')}</p>
-                  </div>
-                ) : (
-                  <label className="border-2 border-dashed border-white/20 rounded-lg p-4 flex flex-col items-center justify-center text-center gap-2 cursor-pointer hover:border-purple-400 transition-all">
-                    <UploadCloud size={20} className="text-gray-400" />
-                    <span className="text-xs text-gray-400">{t('upload.config.referenceUpload')}</span>
-                    <input
-                      type="file"
-                      accept="image/*"
-                      className="hidden"
-                      onChange={handleReferenceImageChange}
-                    />
-                  </label>
-                )}
-              </div>
-            )}
-          </div>
-
-          <button 
-            onClick={handleUploadAndParse} 
-            disabled={(uploadMode === 'file' && !selectedFile) || ((uploadMode === 'text' || uploadMode === 'topic') && !textContent.trim()) || isUploading} 
-            className="w-full py-3 rounded-lg bg-gradient-to-r from-purple-600 to-pink-600 hover:from-purple-700 hover:to-pink-700 disabled:from-gray-600 disabled:to-gray-700 text-white font-semibold flex items-center justify-center gap-2 transition-all"
-          >
-            {isUploading ? (
-              <><Loader2 size={18} className="animate-spin" /> {uploadMode === 'topic' ? t('upload.config.startButton.researching') : t('upload.config.startButton.parsing')}</>
-            ) : (
-              <><ArrowRight size={18} /> {uploadMode === 'topic' ? t('upload.config.startButton.research') : t('upload.config.startButton.parse')}</>
-            )}
-          </button>
-
-          <div className="flex items-start gap-2 text-xs text-gray-500 mt-3 px-1">
-            <Info size={14} className="mt-0.5 text-gray-400 flex-shrink-0" />
-            <p>{t('upload.config.tip')}</p>
-          </div>
-
-          {isUploading && (
-            <div className="mt-4 animate-in fade-in slide-in-from-top-2">
-              <div className="flex justify-between text-xs text-gray-400 mb-1">
-                <span>{progressStatus}</span>
-                <span>{Math.round(progress)}%</span>
-              </div>
-              <div className="h-1.5 bg-white/10 rounded-full overflow-hidden">
-                <div 
-                  className="h-full bg-gradient-to-r from-purple-500 to-pink-500 transition-all duration-300 ease-out"
-                  style={{ width: `${progress}%` }}
-                />
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-
-      {isValidating && (
-        <div className="mt-4 flex items-center gap-2 text-sm text-blue-300 bg-blue-500/10 border border-blue-500/40 rounded-lg px-4 py-3 animate-pulse">
-            <Loader2 size={16} className="animate-spin" />
-            <p>正在验证 API Key 有效性...</p>
-        </div>
-      )}
-
-      {error && (
-        <div className="mt-4 flex items-center gap-2 text-sm text-red-300 bg-red-500/10 border border-red-500/40 rounded-lg px-4 py-3">
-          <AlertCircle size={16} /> {error}
-        </div>
-      )}
-
-      {/* 示例区 */}
-      <div className="space-y-4 mt-8">
-        <div className="flex items-center justify-between flex-wrap gap-2">
-            <div className="flex items-center gap-3">
-            <h3 className="text-sm font-medium text-gray-200">{t('upload.demo.title')}</h3>
-            <a
-              href="https://wcny4qa9krto.feishu.cn/wiki/VXKiwYndwiWAVmkFU6kcqsTenWh"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="group relative inline-flex items-center gap-2 px-3 py-1 rounded-full bg-black/50 border border-white/10 text-xs font-medium text-white overflow-hidden transition-all hover:border-white/30 hover:shadow-[0_0_15px_rgba(168,85,247,0.5)]"
-            >
-              <div className="absolute inset-0 bg-gradient-to-r from-blue-500/20 via-purple-500/20 to-pink-500/20 opacity-0 group-hover:opacity-100 transition-opacity" />
-              <Sparkles size={12} className="text-yellow-300 animate-pulse" />
-              <span className="bg-gradient-to-r from-blue-300 via-purple-300 to-pink-300 bg-clip-text text-transparent group-hover:from-blue-200 group-hover:via-purple-200 group-hover:to-pink-200">
-                {t('upload.demo.more')}
-              </span>
-            </a>
-          </div>
-          <span className="text-[11px] text-gray-500">
-            {t('upload.demo.desc')}
-          </span>
-        </div>
-
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-4 text-xs">
-          <DemoCard
-            title={t('upload.demo.card1.title')}
-            desc={t('upload.demo.card1.desc')}
-            inputImg="/paper2ppt/input_1.png"
-            outputImg="/paper2ppt/ouput_1.png"
-          />
-          <DemoCard
-            title={t('upload.demo.card2.title')}
-            desc={t('upload.demo.card2.desc')}
-            inputImg="/paper2ppt/input_3.png"
-            outputImg="/paper2ppt/ouput_3.png"
-          />
-          <DemoCard
-            title={t('upload.demo.card3.title')}
-            desc={t('upload.demo.card3.desc')}
-            inputImg="/paper2ppt/input_2.png"
-            outputImg="/paper2ppt/ouput_2.png"
-          />
-          <DemoCard
-            title={t('upload.demo.card4.title')}
-            desc={t('upload.demo.card4.desc')}
-            inputImg="/paper2ppt/input_4.png"
-            outputImg="/paper2ppt/ouput_4.png"
-          />
-        </div>
-      </div>
-    </div>
-  );
-
-  const renderOutlineStep = () => (
-    <div className="max-w-5xl mx-auto">
-      <div className="text-center mb-8">
-        <h2 className="text-2xl font-bold text-white mb-2">确认大纲</h2>
-        <p className="text-gray-400">检查从论文提取的内容结构，可编辑、排序或删除</p>
-      </div>
-
-      <div className="glass rounded-xl border border-white/10 p-6 mb-6">
-        <div className="space-y-3">
-          {outlineData.map((slide, index) => (
-            <div 
-              key={slide.id} 
-              className={`flex items-start gap-4 p-4 rounded-lg border transition-all ${
-                editingId === slide.id 
-                  ? 'bg-purple-500/10 border-purple-500/40' 
-                  : 'bg-white/5 border-white/10 hover:border-white/20'
-              }`}
-            >
-              <div className="flex items-center gap-2 pt-1">
-                <GripVertical size={16} className="text-gray-500" />
-                <span className="w-8 h-8 rounded-full bg-purple-500/20 text-purple-300 text-sm font-medium flex items-center justify-center">
-                  {slide.pageNum}
-                </span>
-              </div>
-              
-              <div className="flex-1">
-                {editingId === slide.id ? (
-                  <div className="space-y-3">
-                    <input type="text" value={editContent.title} onChange={e => setEditContent(p => ({ ...p, title: e.target.value }))} className="w-full px-3 py-2 rounded-lg bg-black/40 border border-white/20 text-white text-sm outline-none focus:ring-2 focus:ring-purple-500" placeholder="标题" />
-                    <textarea value={editContent.layout_description} onChange={e => setEditContent(p => ({ ...p, layout_description: e.target.value }))} rows={2} className="w-full px-3 py-2 rounded-lg bg-black/40 border border-white/20 text-white text-sm outline-none focus:ring-2 focus:ring-purple-500 resize-none" placeholder="布局描述" />
-                    <div className="space-y-2">
-                      {editContent.key_points.map((p, i) => (
-                        <div key={i} className="flex gap-2">
-                          <input type="text" value={p} onChange={e => handleKeyPointChange(i, e.target.value)} className="flex-1 px-3 py-2 rounded-lg bg-black/40 border border-white/20 text-white text-sm" placeholder={`要点 ${i + 1}`} />
-                          <button onClick={() => handleRemoveKeyPoint(i)} className="p-2 text-gray-400 hover:text-red-400"><Trash2 size={14} /></button>
-                        </div>
-                      ))}
-                      <button onClick={handleAddKeyPoint} className="px-3 py-1.5 rounded-lg bg-white/5 border border-dashed border-white/20 text-gray-400 text-sm w-full hover:text-purple-400 hover:border-purple-400">+ 添加要点</button>
-                    </div>
-                    <div className="flex gap-2 pt-2">
-                      <button onClick={handleEditSave} className="px-3 py-1.5 rounded-lg bg-purple-500 text-white text-sm flex items-center gap-1"><Check size={14} /> 保存</button>
-                      <button onClick={handleEditCancel} className="px-3 py-1.5 rounded-lg bg-white/10 text-gray-300 text-sm">取消</button>
-                    </div>
-                  </div>
-                ) : (
-                  <>
-                    <div className="mb-2"><h4 className="text-white font-medium">{slide.title}</h4></div>
-                    <p className="text-xs text-purple-400/70 mb-2 italic">📐 {slide.layout_description}</p>
-                    <ul className="space-y-1">
-                      {slide.key_points.map((p, i) => (
-                        <li key={i} className="text-sm text-gray-400 flex items-start gap-2">
-                          <span className="text-purple-400 mt-0.5">•</span><span>{p}</span>
-                        </li>
-                      ))}
-                    </ul>
-                  </>
-                )}
-              </div>
-
-              {editingId !== slide.id && (
-                <div className="flex items-center gap-1">
-                  <button onClick={() => handleMoveSlide(index, 'up')} disabled={index === 0} className="p-2 text-gray-400 hover:text-white disabled:opacity-30"><ChevronUp size={16} /></button>
-                  <button onClick={() => handleMoveSlide(index, 'down')} disabled={index === outlineData.length - 1} className="p-2 text-gray-400 hover:text-white disabled:opacity-30"><ChevronDown size={16} /></button>
-                  <button onClick={() => handleEditStart(slide)} className="p-2 text-gray-400 hover:text-purple-400"><Edit3 size={16} /></button>
-                  <button onClick={() => handleDeleteSlide(slide.id)} className="p-2 text-gray-400 hover:text-red-400"><Trash2 size={16} /></button>
-                </div>
-              )}
-            </div>
-          ))}
-        </div>
-      </div>
-
-      <div className="flex justify-between">
-        <button onClick={() => setCurrentStep('upload')} className="px-6 py-2.5 rounded-lg border border-white/20 text-gray-300 hover:bg-white/10 flex items-center gap-2">
-          <ArrowLeft size={18} /> 返回上传
-        </button>
-        <button onClick={handleConfirmOutline} className="px-6 py-2.5 rounded-lg bg-gradient-to-r from-purple-600 to-pink-600 text-white font-semibold flex items-center gap-2 transition-all">
-          确认并开始生成 <ArrowRight size={18} />
-        </button>
-      </div>
-
-      {error && (
-        <div className="mt-4 flex items-center gap-2 text-sm text-red-300 bg-red-500/10 border border-red-500/40 rounded-lg px-4 py-3">
-          <AlertCircle size={16} /> {error}
-        </div>
-      )}
-    </div>
-  );
-
-  const renderGenerateStep = () => {
-    const currentSlide = outlineData[currentSlideIndex];
-    const currentResult = generateResults[currentSlideIndex];
-    
-    return (
-      <div className="max-w-6xl mx-auto">
-        <div className="text-center mb-6">
-          <h2 className="text-2xl font-bold text-white mb-2">逐页生成</h2>
-          <p className="text-gray-400">第 {currentSlideIndex + 1} / {outlineData.length} 页：{currentSlide?.title}</p>
-        </div>
-
-        <div className="mb-6">
-          <div className="flex gap-1">
-            {generateResults.map((result, index) => (
-              <div key={result.slideId} className={`flex-1 h-2 rounded-full transition-all ${
-                result.status === 'done' ? 'bg-purple-400' : result.status === 'processing' ? 'bg-gradient-to-r from-purple-400 to-pink-400 animate-pulse' : index === currentSlideIndex ? 'bg-purple-400/50' : 'bg-white/10'
-              }`} />
-            ))}
-          </div>
-        </div>
-
-        {currentSlide && (
-          <div className="glass rounded-xl border border-white/10 p-4 mb-4">
-            <div className="mb-3">
-              <h4 className="text-sm text-gray-400 mb-2 flex items-center gap-2"><FileText size={14} className="text-purple-400" /> 布局描述</h4>
-              <p className="text-xs text-purple-400/80 italic">{currentSlide.layout_description}</p>
-            </div>
-            <div className="pt-3 border-t border-white/10">
-              <h4 className="text-sm text-gray-400 mb-2">要点内容</h4>
-              <ul className="grid grid-cols-1 md:grid-cols-2 gap-1">
-                {currentSlide.key_points.slice(0, 4).map((point, idx) => (
-                  <li key={idx} className="text-xs text-gray-400 flex items-start gap-1"><span className="text-purple-400">•</span><span className="line-clamp-1">{point}</span></li>
-                ))}
-                {currentSlide.key_points.length > 4 && (<li className="text-xs text-gray-500 italic">...还有 {currentSlide.key_points.length - 4} 条</li>)}
-              </ul>
-            </div>
-          </div>
-        )}
-
-        <div className="glass rounded-xl border border-white/10 p-6 mb-6">
-          <div className="max-w-3xl mx-auto">
-            <h4 className="text-sm text-gray-400 mb-3 flex items-center justify-center gap-2"><Sparkles size={14} className="text-purple-400" /> AI 生成结果</h4>
-            <div className="rounded-lg overflow-hidden border border-purple-500/30 aspect-[16/9] bg-gradient-to-br from-purple-500/10 to-pink-500/10 flex items-center justify-center">
-              {isGenerating ? (
-                <div className="text-center">
-                  <Loader2 size={40} className="text-purple-400 animate-spin mx-auto mb-3" />
-                  <p className="text-base text-purple-300">{generateResults.every(r => r.status === 'processing') ? '正在批量生成所有页面...' : '正在重新生成当前页...'}</p>
-                  <p className="text-xs text-gray-500 mt-1">{generateResults.every(r => r.status === 'processing') ? `共 ${outlineData.length} 页，请稍候` : 'AI 正在根据您的提示重新创建'}</p>
-                </div>
-              ) : currentResult?.afterImage ? (
-                <img src={currentResult.afterImage} alt="Generated" className="w-full h-full object-contain" />
-              ) : (
-                <div className="text-center"><FileText size={32} className="text-gray-500 mx-auto mb-2" /><span className="text-gray-500">等待生成</span></div>
-              )}
-            </div>
-          </div>
-        </div>
-
-        <div className="glass rounded-xl border border-white/10 p-4 mb-6">
-          <div className="flex items-center gap-3">
-            <MessageSquare size={18} className="text-purple-400" />
-            <input type="text" value={slidePrompt} onChange={e => setSlidePrompt(e.target.value)} placeholder="输入微调 Prompt，然后点击重新生成..." className="flex-1 bg-transparent outline-none text-white text-sm placeholder:text-gray-500" />
-            <button onClick={handleRegenerateSlide} disabled={isGenerating || !slidePrompt.trim()} className="px-4 py-2 rounded-lg bg-white/10 hover:bg-white/20 text-gray-300 text-sm flex items-center gap-2 disabled:opacity-50">
-              <RefreshCw size={14} /> 重新生成
-            </button>
-          </div>
-        </div>
-
-        <div className="flex justify-between">
-          <button onClick={() => setCurrentStep('outline')} className="px-6 py-2.5 rounded-lg border border-white/20 text-gray-300 hover:bg-white/10 flex items-center gap-2">
-            <ArrowLeft size={18} /> 返回大纲
-          </button>
-          <div className="flex gap-3">
-            <button 
-              onClick={() => {
-                if (currentSlideIndex > 0) {
-                  setCurrentSlideIndex(currentSlideIndex - 1);
-                  setSlidePrompt('');
-                }
-              }}
-              disabled={currentSlideIndex === 0 || isGenerating}
-              className="px-6 py-2.5 rounded-lg border border-white/20 text-gray-300 hover:bg-white/10 flex items-center gap-2 disabled:opacity-30"
-            >
-              <ArrowLeft size={18} /> 上一页
-            </button>
-            <button onClick={handleConfirmSlide} disabled={isGenerating || currentResult?.status !== 'done'} className="px-6 py-2.5 rounded-lg bg-gradient-to-r from-purple-600 to-pink-600 text-white font-semibold flex items-center gap-2 disabled:opacity-50">
-              <CheckCircle2 size={18} /> {currentSlideIndex < outlineData.length - 1 ? '确认并继续' : '完成生成'}
-            </button>
-          </div>
-        </div>
-
-        {error && (
-          <div className="mt-4 flex items-center gap-2 text-sm text-red-300 bg-red-500/10 border border-red-500/40 rounded-lg px-4 py-3">
-            <AlertCircle size={16} /> {error}
-          </div>
-        )}
-      </div>
-    );
-  };
-
-  const renderCompleteStep = () => {
-    const doneCount = generateResults.filter(r => r.status === 'done').length;
-    
-    return (
-      <div className="max-w-2xl mx-auto text-center">
-        <div className="mb-8">
-          <div className="w-20 h-20 rounded-full bg-gradient-to-br from-purple-500 to-pink-500 flex items-center justify-center mx-auto mb-4">
-            <CheckCircle2 size={40} className="text-white" />
-          </div>
-          <h2 className="text-2xl font-bold text-white mb-2">生成完成！</h2>
-          <p className="text-gray-400">共处理 {outlineData.length} 页，成功生成 {doneCount} 页</p>
-        </div>
-
-        <div className="glass rounded-xl border border-white/10 p-6 mb-6">
-          <h3 className="text-white font-semibold mb-4">生成结果预览</h3>
-          <div className="grid grid-cols-4 gap-2">
-            {generateResults.map((result, index) => (
-              <div key={result.slideId} className="aspect-[16/9] rounded-lg border border-white/20 overflow-hidden bg-white/5">
-                {result.afterImage ? (
-                  <img src={result.afterImage} alt={`Page ${index + 1}`} className="w-full h-full object-contain" />
-                ) : (
-                  <div className="w-full h-full flex items-center justify-center text-gray-500 text-xs">第 {index + 1} 页</div>
-                )}
-              </div>
-            ))}
-          </div>
-        </div>
-
-        {!(downloadUrl || pdfPreviewUrl) ? (
-          <button onClick={handleGenerateFinal} disabled={isGeneratingFinal} className="px-8 py-3 rounded-lg bg-gradient-to-r from-purple-600 to-pink-600 text-white font-semibold flex items-center justify-center gap-2 mx-auto transition-all">
-            {isGeneratingFinal ? (<><Loader2 size={18} className="animate-spin" /> 正在生成最终文件...</>) : (<><Sparkles size={18} /> 生成最终文件</>)}
-          </button>
-        ) : (
-          <div className="space-y-4">
-            <div className="flex gap-4 justify-center">
-              {/* 已移除 PPTX 下载按钮 */}
-              {pdfPreviewUrl && (
-                <button onClick={handleDownloadPdf} className="px-6 py-3 rounded-lg bg-gradient-to-r from-purple-500 to-pink-500 text-white font-semibold flex items-center gap-2 transition-all">
-                  <Download size={18} /> 下载 PDF
-                </button>
-              )}
-            </div>
-            
-            {/* 引导去 PDF2PPT */}
-            <div className="text-center text-sm text-gray-400 bg-white/5 border border-white/10 rounded-lg p-3">
-              如果需要继续 PDF 转可编辑 PPTX，请前往 <a href="/pdf2ppt" className="text-purple-400 hover:text-purple-300 hover:underline font-medium transition-colors">PDF2PPT 页面</a>
-            </div>
-
-            <div>
-              <button onClick={() => { setCurrentStep('upload'); setSelectedFile(null); setOutlineData([]); setGenerateResults([]); setDownloadUrl(null); setPdfPreviewUrl(null); setResultPath(null); setError(null); }} className="text-sm text-gray-400 hover:text-white transition-colors">
-                <RotateCcw size={14} className="inline mr-1" /> 处理新的论文
-              </button>
-            </div>
-          </div>
-        )}
-
-        {error && (
-          <div className="mt-4 flex items-center gap-2 text-sm text-red-300 bg-red-500/10 border border-red-500/40 rounded-lg px-4 py-3 justify-center">
-            <AlertCircle size={16} /> {error}
-          </div>
-        )}
-
-        {/* 分享与交流群区域 */}
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-4 mt-8 text-left">
-          {/* 获取免费 Key */}
-          <div className="glass rounded-xl border border-white/10 p-5 flex flex-col items-center text-center hover:bg-white/5 transition-colors">
-            <div className="w-12 h-12 rounded-full bg-yellow-500/20 text-yellow-300 flex items-center justify-center mb-3">
-              <Star size={24} />
-            </div>
-            <h4 className="text-white font-semibold mb-2">获取免费 API Key</h4>
-            <p className="text-xs text-gray-400 mb-4 leading-relaxed">
-              点击下方平台图标复制推广文案<br/>
-              分享至朋友圈/小红书/推特，截图联系微信群管理员领 Key！
-            </p>
-            
-            {/* 分享按钮组 */}
-            <div className="flex items-center justify-center gap-4 mb-5 w-full">
-              <button onClick={handleCopyShareText} className="flex flex-col items-center gap-1 group">
-                <div className="w-10 h-10 rounded-full bg-[#00C300]/20 text-[#00C300] flex items-center justify-center border border-[#00C300]/30 group-hover:scale-110 transition-transform">
-                  <MessageSquare size={18} />
-                </div>
-                <span className="text-[10px] text-gray-400">微信</span>
-              </button>
-              <button onClick={handleCopyShareText} className="flex flex-col items-center gap-1 group">
-                <div className="w-10 h-10 rounded-full bg-[#FF2442]/20 text-[#FF2442] flex items-center justify-center border border-[#FF2442]/30 group-hover:scale-110 transition-transform">
-                  <span className="font-bold text-xs">小红书</span>
-                </div>
-                <span className="text-[10px] text-gray-400">小红书</span>
-              </button>
-              <button onClick={handleCopyShareText} className="flex flex-col items-center gap-1 group">
-                <div className="w-10 h-10 rounded-full bg-white/10 text-white flex items-center justify-center border border-white/20 group-hover:scale-110 transition-transform">
-                  <span className="font-bold text-lg">𝕏</span>
-                </div>
-                <span className="text-[10px] text-gray-400">Twitter</span>
-              </button>
-              <button onClick={handleCopyShareText} className="flex flex-col items-center gap-1 group">
-                <div className="w-10 h-10 rounded-full bg-purple-500/20 text-purple-300 flex items-center justify-center border border-purple-500/30 group-hover:scale-110 transition-transform">
-                  <Copy size={18} />
-                </div>
-                <span className="text-[10px] text-gray-400">复制</span>
-              </button>
-            </div>
-
-            {copySuccess && (
-              <div className="mb-4 px-3 py-1 bg-green-500/20 text-green-300 text-xs rounded-full animate-in fade-in zoom-in">
-                ✨ {copySuccess}
-              </div>
-            )}
-
-            <div className="w-full space-y-2">
-               <a href="https://github.com/OpenDCAI/Paper2Any" target="_blank" rel="noopener noreferrer" className="block w-full py-1.5 px-3 rounded bg-white/5 hover:bg-white/10 text-xs text-purple-300 truncate transition-colors border border-white/5 text-center">
-                 ✨如果本项目对你有帮助，可以点个star嘛～
-               </a>
-               <div className="flex gap-2">
-                 <a href="https://github.com/OpenDCAI/Paper2Any" target="_blank" rel="noopener noreferrer" className="flex-1 inline-flex items-center justify-center gap-1 px-2 py-1.5 bg-white/95 hover:bg-white text-gray-900 rounded-full text-[10px] font-semibold transition-all hover:scale-105 shadow-lg">
-                   <Github size={10} />
-                   <span>Agent</span>
-                   <span className="bg-gray-200 text-gray-800 px-1 py-0.5 rounded-full text-[9px] flex items-center gap-0.5"><Star size={7} fill="currentColor" /> {stars.agent || 'Star'}</span>
-                 </a>
-                 <a href="https://github.com/OpenDCAI/DataFlow" target="_blank" rel="noopener noreferrer" className="flex-1 inline-flex items-center justify-center gap-1 px-2 py-1.5 bg-white/95 hover:bg-white text-gray-900 rounded-full text-[10px] font-semibold transition-all hover:scale-105 shadow-lg">
-                   <Github size={10} />
-                   <span>Core</span>
-                   <span className="bg-gray-200 text-gray-800 px-1 py-0.5 rounded-full text-[9px] flex items-center gap-0.5"><Star size={7} fill="currentColor" /> {stars.dataflow || 'Star'}</span>
-                 </a>
-               </div>
-            </div>
-          </div>
-
-          {/* 交流群 */}
-          <div className="glass rounded-xl border border-white/10 p-5 flex flex-col items-center text-center hover:bg-white/5 transition-colors">
-            <div className="w-12 h-12 rounded-full bg-green-500/20 text-green-300 flex items-center justify-center mb-3">
-              <MessageSquare size={24} />
-            </div>
-            <h4 className="text-white font-semibold mb-2">加入交流群</h4>
-            <p className="text-xs text-gray-400 mb-4">
-              效果满意？遇到问题？<br/>欢迎扫码加入交流群反馈与讨论
-            </p>
-            <div className="w-32 h-32 bg-white p-1 rounded-lg mb-2">
-              <img src="/wechat.png" alt="交流群二维码" className="w-full h-full object-contain" />
-            </div>
-            <p className="text-[10px] text-gray-500">扫码加入微信交流群</p>
-          </div>
-        </div>
-      </div>
-    );
-  };
-
-  return (
-    <div className="w-full h-screen flex flex-col bg-[#050512] overflow-hidden">
-      {showBanner && (
-        <div className="w-full bg-gradient-to-r from-purple-600 via-pink-600 to-orange-500 relative overflow-hidden flex-shrink-0">
-          <div className="absolute inset-0 bg-black opacity-20"></div>
-          <div className="absolute inset-0 animate-pulse">
-            <div className="absolute top-0 left-0 w-full h-full bg-gradient-to-r from-transparent via-white to-transparent opacity-10 animate-shimmer"></div>
-          </div>
-          
-          <div className="relative max-w-7xl mx-auto px-4 py-3 flex flex-col sm:flex-row items-center justify-between gap-3">
-            <div className="flex items-center gap-3 flex-wrap justify-center sm:justify-start">
-              <a
-                href="https://github.com/OpenDCAI"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="flex items-center gap-2 bg-white/20 backdrop-blur-sm rounded-full px-3 py-1 hover:bg-white/30 transition-colors"
-              >
-                <Star size={16} className="text-yellow-300 fill-yellow-300 animate-pulse" />
-                <span className="text-xs font-bold text-white">{t('app.githubProject', { ns: 'common' })}</span>
-              </a>
-              
-              <span className="text-sm font-medium text-white">
-                {t('app.exploreMore', { ns: 'common' })}
-              </span>
-            </div>
-
-            <div className="flex items-center gap-2 flex-wrap justify-center">
-              <a
-                href="https://github.com/OpenDCAI/DataFlow"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="inline-flex items-center gap-2 px-4 py-1.5 bg-white/95 hover:bg-white text-gray-900 rounded-full text-xs font-semibold transition-all hover:scale-105 shadow-lg"
-              >
-                <Github size={14} />
-                <span>DataFlow</span>
-                <span className="bg-gray-200 text-gray-800 px-1.5 py-0.5 rounded-full text-[10px] flex items-center gap-0.5"><Star size={8} fill="currentColor" /> {stars.dataflow || 'Star'}</span>
-                <span className="bg-purple-600 text-white px-2 py-0.5 rounded-full text-[10px]">HOT</span>
-              </a>
-
-              <a
-                href="https://github.com/OpenDCAI/Paper2Any"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="inline-flex items-center gap-2 px-4 py-1.5 bg-white/95 hover:bg-white text-gray-900 rounded-full text-xs font-semibold transition-all hover:scale-105 shadow-lg"
-              >
-                <Github size={14} />
-                <span>Paper2Any</span>
-                <span className="bg-gray-200 text-gray-800 px-1.5 py-0.5 rounded-full text-[10px] flex items-center gap-0.5"><Star size={8} fill="currentColor" /> {stars.agent || 'Star'}</span>
-                <span className="bg-pink-600 text-white px-2 py-0.5 rounded-full text-[10px]">NEW</span>
-              </a>
-
-              <a
-                href="https://github.com/OpenDCAI/DataFlex"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="inline-flex items-center gap-2 px-4 py-1.5 bg-white/95 hover:bg-white text-gray-900 rounded-full text-xs font-semibold transition-all hover:scale-105 shadow-lg"
-              >
-                <Github size={14} />
-                <span>DataFlex</span>
-                <span className="bg-gray-200 text-gray-800 px-1.5 py-0.5 rounded-full text-[10px] flex items-center gap-0.5"><Star size={8} fill="currentColor" /> {stars.dataflex || 'Star'}</span>
-                <span className="bg-sky-600 text-white px-2 py-0.5 rounded-full text-[10px]">NEW</span>
-              </a>
-
-              <button
-                onClick={() => setShowBanner(false)}
-                className="p-1 hover:bg-white/20 rounded-full transition-colors"
-                aria-label="关闭"
-              >
-                <X size={16} className="text-white" />
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      <div className="flex-1 overflow-auto">
-        <div className="max-w-7xl mx-auto px-6 py-8 pb-24">
-          {renderStepIndicator()}
-          {currentStep === 'upload' && renderUploadStep()}
-          {currentStep === 'outline' && renderOutlineStep()}
-          {currentStep === 'generate' && renderGenerateStep()}
-          {currentStep === 'complete' && renderCompleteStep()}
-        </div>
-      </div>
-
-      <style>{`
-        @keyframes shimmer {
-          0% { transform: translateX(-100%); }
-          100% { transform: translateX(100%); }
-        }
-        .animate-shimmer {
-          animation: shimmer 3s infinite;
-        }
-        .animate-shimmer-fast {
-          animation: shimmer 1.5s infinite;
-        }
-        .glass { background: rgba(255, 255, 255, 0.03); backdrop-filter: blur(10px); }
-        .demo-input-placeholder {
-          min-height: 80px;
-        }
-        .demo-output-placeholder {
-          min-height: 80px;
-        }
-      `}</style>
-    </div>
-  );
-};
-
-interface DemoCardProps {
-  title: string;
-  desc: string;
-  inputImg?: string;
-  outputImg?: string;
-}
-
-const DemoCard = ({ title, desc, inputImg, outputImg }: DemoCardProps) => {
-  return (
-    <div className="glass rounded-lg border border-white/10 p-3 flex flex-col gap-2 hover:bg-white/5 transition-colors">
-      <div className="flex gap-2">
-        {/* 左侧：输入示例图片 */}
-        <div className="flex-1 rounded-md bg-white/5 border border-dashed border-white/10 flex items-center justify-center demo-input-placeholder overflow-hidden">
-          {inputImg ? (
-            <img
-              src={inputImg}
-              alt="输入示例图"
-              className="w-full h-full object-cover"
-            />
-          ) : (
-            <span className="text-[10px] text-gray-400">输入示例图（待替换）</span>
-          )}
-        </div>
-        {/* 右侧：输出 PPTX 示例图片 */}
-        <div className="flex-1 rounded-md bg-primary-500/10 border border-dashed border-primary-300/40 flex items-center justify-center demo-output-placeholder overflow-hidden">
-          {outputImg ? (
-            <img
-              src={outputImg}
-              alt="PPTX 示例图"
-              className="w-full h-full object-cover"
-            />
-          ) : (
-            <span className="text-[10px] text-primary-200">PPTX 示例图（待替换）</span>
-          )}
-        </div>
-      </div>
-      <div>
-        <p className="text-[13px] text-white font-medium mb-1">{title}</p>
-        <p className="text-[11px] text-gray-400 leading-snug">{desc}</p>
-      </div>
-    </div>
-  );
-};
+import Paper2PptPage from './paper2ppt';
 
 export default Paper2PptPage;
