+++ frontend-workflow/src/components/Pdf2PptPage.tsx
@@ -12,6 +12,7 @@ import { useAuthStore } from '../stores/authStore';
 import QRCodeTooltip from './QRCodeTooltip';
 
 const MAX_FILE_SIZE = 50 * 1024 * 1024; // 50MB
+const STORAGE_KEY = 'pdf2ppt-storage';
 
 // ============== 主组件 ==============
 const Pdf2PptPage = () => {
@@ -105,6 +106,41 @@ const Pdf2PptPage = () => {
   const [llmApiUrl, setLlmApiUrl] = useState('https://api.apiyi.com/v1');
   const [apiKey, setApiKey] = useState('');
   const [genFigModel, setGenFigModel] = useState('gemini-3-pro-image-preview');
+
+  // 从 localStorage 恢复配置
+  useEffect(() => {
+    if (typeof window === 'undefined') return;
+    try {
+      const raw = localStorage.getItem(STORAGE_KEY);
+      if (!raw) return;
+      const saved = JSON.parse(raw);
+      
+      if (saved.inviteCode) setInviteCode(saved.inviteCode);
+      if (saved.useAiEdit !== undefined) setUseAiEdit(saved.useAiEdit);
+      if (saved.llmApiUrl) setLlmApiUrl(saved.llmApiUrl);
+      if (saved.apiKey) setApiKey(saved.apiKey);
+      if (saved.genFigModel) setGenFigModel(saved.genFigModel);
+    } catch (e) {
+      console.error('Failed to restore pdf2ppt config', e);
+    }
+  }, []);
+
+  // 将配置写入 localStorage
+  useEffect(() => {
+    if (typeof window === 'undefined') return;
+    const data = {
+      inviteCode,
+      useAiEdit,
+      llmApiUrl,
+      apiKey,
+      genFigModel,
+    };
+    try {
+      localStorage.setItem(STORAGE_KEY, JSON.stringify(data));
+    } catch (e) {
+      console.error('Failed to persist pdf2ppt config', e);
+    }
+  }, [inviteCode, useAiEdit, llmApiUrl, apiKey, genFigModel]);
 
   const validateDocFile = (file: File): boolean => {
     const ext = file.name.split('.').pop()?.toLowerCase();
@@ -559,7 +595,7 @@ const Pdf2PptPage = () => {
 
                 <div className="flex items-start gap-2 text-xs text-gray-500 mt-3 px-1">
                   <Info size={14} className="mt-0.5 text-gray-400 flex-shrink-0" />
-                  <p>{t('errors.serverBusy')}</p>
+                  <p>处理时间取决于文件大小，请耐心等待。</p>
                 </div>
               </>
             ) : (
