+++ frontend-workflow/src/components/paper2ppt/DemoCard.tsx
@@ -1,0 +1,47 @@
+import React from 'react';
+
+interface DemoCardProps {
+  title: string;
+  desc: string;
+  inputImg?: string;
+  outputImg?: string;
+}
+
+const DemoCard: React.FC<DemoCardProps> = ({ title, desc, inputImg, outputImg }) => {
+  return (
+    <div className="glass rounded-lg border border-white/10 p-3 flex flex-col gap-2 hover:bg-white/5 transition-colors">
+      <div className="flex gap-2">
+        {/* 左侧：输入示例图片 */}
+        <div className="flex-1 rounded-md bg-white/5 border border-dashed border-white/10 flex items-center justify-center demo-input-placeholder overflow-hidden">
+          {inputImg ? (
+            <img
+              src={inputImg}
+              alt="输入示例图"
+              className="w-full h-full object-cover"
+            />
+          ) : (
+            <span className="text-[10px] text-gray-400">输入示例图（待替换）</span>
+          )}
+        </div>
+        {/* 右侧：输出 PPTX 示例图片 */}
+        <div className="flex-1 rounded-md bg-primary-500/10 border border-dashed border-primary-300/40 flex items-center justify-center demo-output-placeholder overflow-hidden">
+          {outputImg ? (
+            <img
+              src={outputImg}
+              alt="PPTX 示例图"
+              className="w-full h-full object-cover"
+            />
+          ) : (
+            <span className="text-[10px] text-primary-200">PPTX 示例图（待替换）</span>
+          )}
+        </div>
+      </div>
+      <div>
+        <p className="text-[13px] text-white font-medium mb-1">{title}</p>
+        <p className="text-[11px] text-gray-400 leading-snug">{desc}</p>
+      </div>
+    </div>
+  );
+};
+
+export default DemoCard;
