+++ frontend-workflow/src/components/paper2ppt/constants.ts
@@ -1,0 +1,9 @@
+export const MAX_FILE_SIZE = 50 * 1024 * 1024; // 50MB
+export const STORAGE_KEY = 'paper2ppt-storage';
+
+export const STYLE_PRESETS: Record<string, string> = {
+  modern: '现代简约风格，使用干净的线条和充足的留白',
+  business: '商务专业风格，稳重大气，适合企业演示',
+  academic: '学术报告风格，清晰的层次结构，适合论文汇报',
+  creative: '创意设计风格，活泼生动，色彩丰富',
+};
